requirements.txt
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Compress==1.14
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.22
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.2.3